Update Control_Image.md
This is my first time submitting a pull request in Github so I am keeping it small. I added the control image for v4 to the landscapes section on this page, which you already had in your save images. I confirmed by also requesting the landscape image for seed 4776 to make sure it was the right one.